Renaming the DynDataDistributer.py file to dyn_data_distributer.py to follow the PEP8 guidelines
